requirements.txt
torch==1.10.2+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html
tensorboard
datasets>=1.5.0
tensorflow>=2.4.0
transformers>=4.4.0
tokenizers<=0.10.2
hydra-core==1.0.6
matplotlib
tqdm
soynlp
seaborn
Flask
Flask_restful
tensorflow_addons
pymysql